+++ src/Raven.CodeAnalysis/Diagnostics/Infrastructure/BuiltInAnalyzers.cs
@@ -26,6 +26,7 @@ public static class BuiltInAnalyzers
         project = AddAnalyzerIfMissing<UnusedParameterAnalyzer>(project);
         project = AddAnalyzerIfMissing<UnusedImportDirectiveAnalyzer>(project);
         project = AddAnalyzerIfMissing<UnhandledMemberReturnValueAnalyzer>(project);
+        project = AddAnalyzerIfMissing<DisposableObjectAnalyzer>(project);
         project = AddAnalyzerIfMissing<ImmutableCollectionOperationResultAnalyzer>(project);
         project = AddAnalyzerIfMissing<PreferDuLinqExtensionsAnalyzer>(project);
         project = AddAnalyzerIfMissing<ConstructorParameterNamingAnalyzer>(project);